requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiohttp-retry==2.9.1
aiosignal==1.4.0
asgiref==3.10.0
async-timeout==5.0.1
attrs==25.4.0
certifi==2025.11.12
charset-normalizer==3.4.4
Django==5.2.7
django-cors-headers==4.9.0
django-environ==0.12.0
django-rest-framework==0.1.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
frozenlist==1.8.0
idna==3.11
multidict==6.7.0
propcache==0.4.1
PyJWT==2.10.1
python-decouple==3.8
requests==2.32.5
sqlparse==0.5.3
twilio==9.9.0
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.6.2
yarl==1.22.0
